requirements.txt
openai-whisper
ffmpeg-python
colorlog
tqdm
